# api/openapi.yaml
openapi: 3.0.3
info:
  title: wireguardd API
  version: 0.1.0
  description: REST API for WireGuard interface and peer management.
servers:
  - url: http://127.0.0.1:51880
security:
  - bearerAuth: []
paths:
...
  /v1/interfaces/{name}:
    parameters:
      - $ref: "#/components/parameters/Name"
    get:
      summary: Get interface
      responses:
        "200":
          description: Interface
    patch:
      summary: Update interface
      responses:
        "200":
          description: Updated
    delete:
      summary: Delete interface
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Name:
      name: name
      in: path
      required: true
      schema:
        type: string
...